Add crossterm dependency and refactor terminal utils
