+++ Cargo.toml
@@ -14,3 +14,4 @@ async-fs = "1.6.0"
 stringmetrics = "2.2.2"
 futures = "0.3.28"
 tokio = { version = "1.28.2", features = ["full"] }
+crossterm = "0.26.1"
